LPS-55885 Moves Robots and Sitemap actions to portal
