+++ portal-impl/src/portal.properties
@@ -3573,9 +3573,6 @@
         /image_gallery_display/find_folder,\
         /image_gallery_display/find_image,\
         \
-        /layouts_admin/robots,\
-        /layouts_admin/sitemap,\
-        \
         /login/facebook_connect_oauth,\
         \
         /message_boards/find_category,\
@@ -3594,8 +3591,10 @@
         /portal/open_id_request,\
         /portal/open_id_response,\
         /portal/portlet_url,\
+		/portal/robots,\
         /portal/session_click,\
         /portal/session_tree_js_click,\
+        /portal/sitemap,\
         /portal/status,\
         \
         /shopping/notify
